kernel: avoid userland namespace pollution

+++ src/init/calls.c
@@ -2,45 +2,47 @@
 
 #include <kern/kobject.h>
 #include <kern/syscall.h>
+#include <stdint.h>
 
 extern long __do_syscall(
-    u64 no, u64 arg1, u64 arg2, u64 arg3, u64 arg4, u64 arg5, u64 arg6);
+    uint64_t no, uint64_t arg1, uint64_t arg2, uint64_t arg3, uint64_t arg4,
+    uint64_t arg5, uint64_t arg6);
 
-s64
+int64_t
 as_create(void)
 {
         return __do_syscall(SYSCALL_AS_CREATE, 0, 0, 0, 0, 0, 0);
 }
 
-s64
+int64_t
 as_clone(kobject_handler_t as)
 {
         return __do_syscall(SYSCALL_AS_CLONE, as, 0, 0, 0, 0, 0);
 }
 
-s64
+int64_t
 as_destroy(kobject_handler_t as)
 {
         return __do_syscall(SYSCALL_AS_DESTROY, as, 0, 0, 0, 0, 0);
 }
 
-long
-mmap(kobject_handler_t as, u64 addr, u64 size, uint flag)
+int64_t
+mmap(kobject_handler_t as, uint64_t addr, uint64_t size, uint64_t flag)
 {
         return __do_syscall(SYSCALL_MMAP, as, addr, size, flag, 0, 0);
 }
 
-long
+int64_t
 mtransfer(
-    kobject_handler_t das, kobject_handler_t sas, uintptr vaddr_d,
-    uintptr vaddr_s, size_t size, uint flag)
+    kobject_handler_t das, kobject_handler_t sas, uintptr_t vaddr_d,
+    uintptr_t vaddr_s, size_t size, uint64_t flag)
 {
         return __do_syscall(
             SYSCALL_MTRANSFER, das, sas, vaddr_d, vaddr_s, size, flag);
 }
 
 long
-munmap(kobject_handler_t as, uintptr addr, size_t size)
+munmap(kobject_handler_t as, uintptr_t addr, size_t size)
 {
         return __do_syscall(SYSCALL_MUNMAP, as, addr, size, 0, 0, 0);
 }
@@ -49,13 +51,14 @@ long
 port_create(char *name, unsigned long namelen)
 {
         return __do_syscall(
-            SYSCALL_PORT_CREATE, (u64)name, namelen, 0, 0, 0, 0);
+            SYSCALL_PORT_CREATE, (uint64_t)name, namelen, 0, 0, 0, 0);
 }
 
 long
 port_open(char *name, unsigned long namelen)
 {
-        return __do_syscall(SYSCALL_PORT_OPEN, (u64)name, namelen, 0, 0, 0, 0);
+        return __do_syscall(
+            SYSCALL_PORT_OPEN, (uint64_t)name, namelen, 0, 0, 0, 0);
 }
 
 long
@@ -68,7 +71,7 @@ long
 port_request(kobject_handler_t handler, port_request_t *req)
 {
         return __do_syscall(
-            SYSCALL_PORT_REQUEST, handler, (u64)req, 0, 0, 0, 0);
+            SYSCALL_PORT_REQUEST, handler, (uint64_t)req, 0, 0, 0, 0);
 }
 
 long
@@ -76,33 +79,35 @@ port_receive(
     kobject_handler_t handler, port_request_t *req, void *buf, size_t buflen)
 {
         return __do_syscall(
-            SYSCALL_PORT_RECEIVE, handler, (u64)req, (u64)buf, buflen, 0, 0);
+            SYSCALL_PORT_RECEIVE, handler, (uint64_t)req, (uint64_t)buf, buflen,
+            0, 0);
 }
 
 long
 port_response(
-    kobject_handler_t handler, u64 retval_small, void *retval,
+    kobject_handler_t handler, uint64_t retval_small, void *retval,
     size_t retval_size)
 {
         return __do_syscall(
-            SYSCALL_PORT_RESPONSE, handler, retval_small, (u64)retval,
+            SYSCALL_PORT_RESPONSE, handler, retval_small, (uint64_t)retval,
             retval_size, 0, 0);
 }
 
 long
-process_spawn(kobject_handler_t as, u64 entry)
+process_spawn(kobject_handler_t as, uint64_t entry)
 {
         return __do_syscall(SYSCALL_PROCESS_SPAWN, as, entry, 0, 0, 0, 0);
 }
 
-s64
+int64_t
 process_wait(pid_t pid, process_state_t *state)
 {
-        return __do_syscall(SYSCALL_PROCESS_WAIT, pid, (u64)state, 0, 0, 0, 0);
+        return __do_syscall(
+            SYSCALL_PROCESS_WAIT, pid, (uint64_t)state, 0, 0, 0, 0);
 }
 
-s64
-process_exit(u64 retval)
+int64_t
+process_exit(uint64_t retval)
 {
         return __do_syscall(SYSCALL_PROCESS_EXIT, retval, 0, 0, 0, 0, 0);
 }
@@ -110,7 +115,7 @@ process_exit(u64 retval)
 void
 print(const char *str)
 {
-        u64 len;
+        uint64_t len;
         for (len = 0; str[len]; ++len)
                 ;
